+++ src/task_3_1_fabric_2/MoreThanOneEquationResult.java
@@ -3,9 +3,9 @@ package task_3_1_fabric_2;
 import java.util.List;
 
 public class MoreThanOneEquationResult implements EquationResult {
-    List<Double> value;
+    List<Float> value;
 
-    MoreThanOneEquationResult(List<Double> value) {
+    MoreThanOneEquationResult(List<Float> value) {
         this.value = value;
     }
 }
